requirements.txt
# Core Data Science Libraries
pandas>=1.5.0
numpy>=1.24.0
scipy>=1.10.0

# Data Visualization
plotly>=5.15.0
seaborn>=0.12.0
matplotlib>=3.6.0

# Machine Learning & Forecasting
scikit-learn>=1.3.0
prophet>=1.1.4
statsmodels>=0.14.0
xgboost>=1.7.0

# Web Framework for Dashboard
streamlit>=1.25.0
dash>=2.12.0
dash-bootstrap-components>=1.4.0

# Data Collection & APIs
requests>=2.31.0
beautifulsoup4>=4.12.0
selenium>=4.10.0
yfinance>=0.2.0
fredapi>=0.5.0
wbdata>=0.3.0
urllib3>=1.26.0
aiohttp>=3.8.0

# Database & Storage
sqlite3
openpyxl>=3.1.0

# Utilities
python-dotenv>=1.0.0
jupyter>=1.0.0
ipywidgets>=8.0.0
tqdm>=4.65.0
